Remove std::ios_base::Init
One test executable in our project was consistently crashing on shutdown due to a nullptr access when destructing that `std::ios_base::Init` instance. I haven't figured out why that happens, but removing this member fixed the problem. It should not be necessary since C++11.

I was using visual studio 15.7.6.